probably finished

+++ Bank/Customer/Controllers/Business/DeleteBusinessCustomer.php
@@ -2,6 +2,8 @@
 
 namespace Bank\Customer\Controllers\Business;
 
+use Bank\Account\CreditAccount;
+use Bank\Account\DepositAccount;
 use Bank\Customer\BussinessCustomer;
 use Bank\Customer\ICustomer;
 use Bank\Services\Persistence\NotFoundException;
@@ -45,17 +47,38 @@ class DeleteBusinessCustomer implements ControllerInterface
     public function execute($request, $response)
     {
         try {
-            $customer = Database::GetEntityManager()->getRepository(BussinessCustomer::class)
+            $this->customer = Database::GetEntityManager()->getRepository(BussinessCustomer::class)
                 ->find($request->id);
+            $this->customer->setAccounts(Database::GetEntityManager()->getRepository(CreditAccount::class)
+                ->findBy(array('customerId' => $this->customer->getId())));
+            $this->customer->setAccounts(Database::GetEntityManager()->getRepository(DepositAccount::class)
+                ->findBy(array('customerId' => $this->customer->getId())));
 
+            if(!$this->customer)
+            {
+                throw new NotFoundException();
+            }
+            if($this->customer->getAccounts()) {
+                foreach ($this->customer->getAccounts() as $account) {
+                    Database::GetEntityManager()->remove($account);
+                }
+            }
 
             Database::GetEntityManager()->remove($this->customer);
             Database::GetEntityManager()->flush();
 
-            return "Customer deleted!";
+            $html = <<<HTML
+<form>
+    <label>Customer Deleted</label>
+	<button type="submit" formaction="/">Home</button>
+</form>
+
+HTML;
+
+            return $html;
 
         } catch (NotFoundException $e) {
-            return "Sorry, the account not found";
+            return $response->redirect("/notfound");
         }
     }
 }
